xplain-hackathon/requirements.txt
aixplain
google_auth_oauthlib
flask
google-generativeai
pytube
google-auth 
google-auth-httplib2   
requests
pymongo